flask_api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.4.0
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
numpy==1.26.4